feat(form): prevent displaying of empty forms
