+++ addon/templates/components/cf-navigation-item.hbs
@@ -1,4 +1,4 @@
-{{#link-to (query-params section=section subSection=_subSection)}}
+{{#link-to (query-params section=section subSection=subSection)}}
   <div class="uk-flex uk-flex-between uk-flex-middle">
     <div>{{label}}</div>
     <span
