+++ addon/templates/components/cf-pagination.hbs
@@ -1,14 +1,14 @@
 <div uk-grid>
   <div class="uk-text-left uk-width-1-2">
-    {{#if previousSection}}
-      {{#link-to (query-params section=previousSection.section subSection=previousSection.subSection) class="uk-button uk-button-default"}}
+    {{#if previous}}
+      {{#link-to (query-params section=previous.section subSection=previous.subSection) class="uk-button uk-button-default"}}
         {{t "caluma.form.navigation.previous"}}
       {{/link-to}}
     {{/if}}
   </div>
   <div class="uk-text-right uk-width-1-2">
-    {{#if nextSection}}
-      {{#link-to (query-params section=nextSection.section subSection=nextSection.subSection) class="uk-button uk-button-default"}}
+    {{#if next}}
+      {{#link-to (query-params section=next.section subSection=next.subSection) class="uk-button uk-button-default"}}
         {{t "caluma.form.navigation.next"}}
       {{/link-to}}
     {{/if}}
